Move gen_rule_consts and gen_token_epp to the codegen module
